3.9.5 Epsilon

+++ source/de/quippy/javamod/multimedia/mod/mixer/BasicModMixer.java
@@ -762,7 +762,7 @@ public abstract class BasicModMixer
 					accurateBuffer = accurateBuffer * swingFactor / (double)(ModConstants.TEMPOSWING_UNITY);
 				}
 				samplesPerTick = (int)(accurateBuffer);
-				bufferDiff += accurateBuffer - samplesPerTick;
+				bufferDiff += accurateBuffer - (double)(samplesPerTick);
 				if (bufferDiff>=1)
 				{
 					samplesPerTick++;
